fix!: adjust the structure of `commit::ancestors` to be more consistent with other plumbing crates.
This makes sure that each time is reached by only a single path.
